Support unmarshalling JSON ordered map (#845)
* Support unmarshalling JSON ordered map

* Use MethodByName helper (#846)

* improve func name
