+++ ytypes/list_test.go
@@ -524,212 +524,6 @@ func TestUnmarshalUnkeyedList(t *testing.T) {
 	}
 }
 
-func TestUnmarshalKeyedList(t *testing.T) {
-	keyListSchema := func() *yang.Entry {
-		return &yang.Entry{
-			Name:     "key-list",
-			Kind:     yang.DirectoryEntry,
-			ListAttr: yang.NewDefaultListAttr(),
-			Key:      "key",
-			Config:   yang.TSTrue,
-			Dir: map[string]*yang.Entry{
-				"key": {
-					Kind: yang.LeafEntry,
-					Name: "key",
-					Type: &yang.YangType{Kind: yang.Ystring},
-				},
-				"leaf-field": {
-					Kind: yang.LeafEntry,
-					Name: "leaf-field",
-					Type: &yang.YangType{Kind: yang.Yint32},
-				},
-				"leaf-field2": {
-					Kind: yang.LeafEntry,
-					Name: "leaf-field2",
-					Type: &yang.YangType{Kind: yang.Yint32},
-				},
-			},
-		}
-	}
-
-	containerWithLeafListSchema := &yang.Entry{
-		Name: "container",
-		Kind: yang.DirectoryEntry,
-		Dir: map[string]*yang.Entry{
-			"key-list": keyListSchema(),
-		},
-	}
-	addParents(containerWithLeafListSchema)
-
-	containerWithPreferConfigSchema := &yang.Entry{
-		Name: "container",
-		Kind: yang.DirectoryEntry,
-		Dir: map[string]*yang.Entry{
-			"config": {
-				Name: "config",
-				Kind: yang.DirectoryEntry,
-				Dir: map[string]*yang.Entry{
-					"key-list": keyListSchema(),
-				},
-			},
-			"state": {
-				Name: "state",
-				Kind: yang.DirectoryEntry,
-				Dir: map[string]*yang.Entry{
-					"key-list": keyListSchema(),
-				},
-			},
-		},
-	}
-	addParents(containerWithPreferConfigSchema)
-
-	type ListElemStruct struct {
-		Key        *string `path:"key"`
-		LeafField  *int32  `path:"leaf-field"`
-		LeafField2 *int32  `path:"leaf-field2"`
-	}
-	type ContainerStruct struct {
-		KeyList map[string]*ListElemStruct `path:"key-list"`
-	}
-
-	type ContainerStructPreferConfig struct {
-		KeyList map[string]*ListElemStruct `path:"config/key-list" shadow-path:"state/key-list"`
-	}
-
-	tests := []struct {
-		desc    string
-		json    string
-		schema  *yang.Entry
-		parent  interface{}
-		want    interface{}
-		opts    []UnmarshalOpt
-		wantErr string
-	}{
-		{
-			desc:   "success",
-			json:   `{ "key-list" : [ { "key" : "forty-two", "leaf-field" : 42} ] }`,
-			schema: containerWithLeafListSchema,
-			parent: &ContainerStruct{},
-			want: &ContainerStruct{
-				KeyList: map[string]*ListElemStruct{
-					"forty-two": {
-						Key:       ygot.String("forty-two"),
-						LeafField: ygot.Int32(42),
-					},
-				},
-			},
-		},
-		{
-			desc:   "success with already-instantiated list element",
-			json:   `{ "key-list" : [ { "key" : "forty-two", "leaf-field" : 42} ] }`,
-			schema: containerWithLeafListSchema,
-			parent: &ContainerStruct{
-				KeyList: map[string]*ListElemStruct{
-					"forty-two": {
-						Key:        ygot.String("forty-two"),
-						LeafField2: ygot.Int32(43),
-					},
-				},
-			},
-			want: &ContainerStruct{
-				KeyList: map[string]*ListElemStruct{
-					"forty-two": {
-						Key:        ygot.String("forty-two"),
-						LeafField:  ygot.Int32(42),
-						LeafField2: ygot.Int32(43),
-					},
-				},
-			},
-		},
-		{
-			desc:   "success with config path",
-			json:   `{ "config": { "key-list" : [ { "key" : "forty-two", "leaf-field" : 42} ] } }`,
-			schema: containerWithPreferConfigSchema,
-			parent: &ContainerStructPreferConfig{},
-			want: &ContainerStructPreferConfig{
-				KeyList: map[string]*ListElemStruct{
-					"forty-two": {
-						Key:       ygot.String("forty-two"),
-						LeafField: ygot.Int32(42),
-					},
-				},
-			},
-		},
-		{
-			desc:   "success ignoring shadowed state path",
-			json:   `{ "state": { "key-list" : [ { "key" : "forty-two", "leaf-field" : 42} ] } }`,
-			schema: containerWithPreferConfigSchema,
-			parent: &ContainerStructPreferConfig{},
-			want:   &ContainerStructPreferConfig{},
-		},
-		{
-			desc:   "success ignoring path with preferShadowPath",
-			json:   `{ "config": { "key-list" : [ { "key" : "forty-two", "leaf-field" : 42} ] } }`,
-			opts:   []UnmarshalOpt{&PreferShadowPath{}},
-			schema: containerWithPreferConfigSchema,
-			parent: &ContainerStructPreferConfig{},
-			want:   &ContainerStructPreferConfig{},
-		},
-		{
-			desc:   "success unmarshalling shadow path",
-			json:   `{ "state": { "key-list" : [ { "key" : "forty-two", "leaf-field" : 42} ] } }`,
-			opts:   []UnmarshalOpt{&PreferShadowPath{}},
-			schema: containerWithPreferConfigSchema,
-			parent: &ContainerStructPreferConfig{},
-			want: &ContainerStructPreferConfig{
-				KeyList: map[string]*ListElemStruct{
-					"forty-two": {
-						Key:       ygot.String("forty-two"),
-						LeafField: ygot.Int32(42),
-					},
-				},
-			},
-		},
-		{
-			desc:    "bad field",
-			json:    `{ "key-list" : [ { "key" : "forty-two", "bad-field" : 42} ] }`,
-			schema:  containerWithLeafListSchema,
-			parent:  &ContainerStruct{},
-			wantErr: `parent container key-list (type *ytypes.ListElemStruct): JSON contains unexpected field bad-field`,
-		},
-		{
-			desc:   "ignore unknown field",
-			json:   `{ "key-list" : [ { "key" : "forty-two", "bad-field" : 42} ] }`,
-			opts:   []UnmarshalOpt{&IgnoreExtraFields{}},
-			schema: containerWithLeafListSchema,
-			parent: &ContainerStruct{},
-			want: &ContainerStruct{
-				KeyList: map[string]*ListElemStruct{
-					"forty-two": {
-						Key: ygot.String("forty-two"),
-					},
-				},
-			},
-		},
-	}
-
-	var jsonTree interface{}
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			if err := json.Unmarshal([]byte(tt.json), &jsonTree); err != nil {
-				t.Fatalf("%s : %s", tt.desc, err)
-			}
-
-			err := Unmarshal(tt.schema, tt.parent, jsonTree, tt.opts...)
-			if got, want := errToString(err), tt.wantErr; got != want {
-				t.Errorf("%s: Unmarshal got error: %v, want error: %v", tt.desc, got, want)
-			}
-			testErrLog(t, tt.desc, err)
-			if err == nil {
-				got, want := tt.parent, tt.want
-				if diff := cmp.Diff(want, got); diff != "" {
-					t.Errorf("%s: Unmarshal (-want, +got):\n%s", tt.desc, diff)
-				}
-			}
-		})
-	}
-}
-
 func TestUnmarshalStructKeyedList(t *testing.T) {
 	containerWithLeafListSchema := &yang.Entry{
 		Name: "container",
@@ -821,84 +615,6 @@ func TestUnmarshalStructKeyedList(t *testing.T) {
 			if err == nil {
 				got, want := parent, tt.want
 				if diff := cmp.Diff(want, got); diff != "" {
-					t.Errorf("%s: Unmarshal (-want, +got):\n%s", tt.desc, diff)
-				}
-			}
-		})
-	}
-}
-
-func TestUnmarshalSingleListElement(t *testing.T) {
-	listSchema := &yang.Entry{
-		Name:     "struct-list",
-		Kind:     yang.DirectoryEntry,
-		ListAttr: yang.NewDefaultListAttr(),
-		Dir: map[string]*yang.Entry{
-			"leaf-field": {
-				Kind: yang.LeafEntry,
-				Name: "leaf-field",
-				Type: &yang.YangType{Kind: yang.Yint32},
-			},
-			"enum-leaf-field": {
-				Kind: yang.LeafEntry,
-				Name: "enum-leaf-field",
-				Type: &yang.YangType{Kind: yang.Yenum},
-			},
-			"leaf2-field": {
-				Kind: yang.LeafEntry,
-				Name: "leaf2-field",
-				Type: &yang.YangType{Kind: yang.Yint64},
-			},
-		},
-	}
-
-	type ListElemStruct struct {
-		LeafName  *int32   `path:"leaf-field"`
-		EnumLeaf  EnumType `path:"enum-leaf-field"`
-		Leaf2Name *int64   `path:"leaf2-field"`
-	}
-
-	tests := []struct {
-		desc     string
-		json     string
-		inParent any
-		want     any
-		wantErr  string
-	}{
-		{
-			desc: "success",
-			inParent: &ListElemStruct{
-				Leaf2Name: ygot.Int64(42),
-			},
-			json: `{ "leaf-field" : 42, "enum-leaf-field" : "E_VALUE_FORTY_TWO"}`,
-			want: &ListElemStruct{
-				LeafName:  ygot.Int32(42),
-				Leaf2Name: ygot.Int64(42),
-				EnumLeaf:  42,
-			},
-		},
-		{
-			desc:     "bad field",
-			inParent: &ListElemStruct{},
-			json:     `{ "leaf-field" : 42, "bad-field" : "E_VALUE_FORTY_TWO"}`,
-			wantErr:  `parent container struct-list (type *ytypes.ListElemStruct): JSON contains unexpected field bad-field`,
-		},
-	}
-
-	var jsonTree interface{}
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			if err := json.Unmarshal([]byte(tt.json), &jsonTree); err != nil {
-				t.Fatalf("%s : %s", tt.desc, err)
-			}
-
-			err := Unmarshal(listSchema, tt.inParent, jsonTree)
-			if got, want := errToString(err), tt.wantErr; got != want {
-				t.Errorf("%s: Unmarshal got error: %v, want error: %v", tt.desc, got, want)
-			}
-			testErrLog(t, tt.desc, err)
-			if err == nil {
-				if diff := cmp.Diff(tt.inParent, tt.want); diff != "" {
 					t.Errorf("%s: Unmarshal (-want, +got):\n%s", tt.desc, diff)
 				}
 			}
